core: support 32 bit system IDs
Groundwork for MAVLink 2's 32 bit system IDs as proposed in
ArduPilot/pymavlink#1229, so that e.g. an IPv4 address can be used
directly as a system ID.

System IDs widen from uint8_t to uint32_t throughout: MavlinkAddress,
Mavsdk::Configuration, System, and every plugin's target_system_id.
Component IDs stay 8 bit. This breaks the C++, C and Python APIs, which
is why it lands before the next major. Mavsdk::MavlinkMessage and the
MavlinkDirect proto were already uint32 and are unchanged.

Two places had to stop truncating:

- Routing. A target above 255 travels in an extended header and the
  payload's target_system field then reads as 0, which
  get_target_system_id() would have reported as a broadcast. It now
  checks IFLAG_TARGETTED first. While widening its return type it also
  turned out that _MAV_PAYLOAD yields a char*, so a target above 127
  sign extended; hence the explicit cast.
- Command acks. mavlink_command_ack_t's target_system is 8 bit, so acks
  are now packed with the originating system ID passed alongside rather
  than encoded from the struct.

Depends on two upstream changes:

- libmavlike is bumped to fbcb13f, which makes the MAVLink 2 header
  variable length (10 to 18 bytes) and adds golden frame tests against
  the C implementation's output.
- The MAVLink C library cannot carry a wide system ID until
  mavlink/mavlink bumps its pymavlink submodule to include the PR
  above. MAVLINK_HASH therefore stays at upstream d6a7eeaf and the
  32 bit paths sit behind #ifdef MAVLINK_IFLAG_TARGETTED, including
  four of the five new system tests. They were verified by generating
  the C headers from the PR branch locally: all 103 system tests and
  413 unit tests pass with them, and 99 + 413 pass without.

+++ cpp/src/mavsdk/core/connection.cpp
@@ -124,7 +124,7 @@ unsigned Connection::forwarding_connections_count()
     return _forwarding_connections_count;
 }
 
-bool Connection::has_system_id(uint8_t system_id)
+bool Connection::has_system_id(uint32_t system_id)
 {
     std::lock_guard<std::mutex> lock(_system_ids_mutex);
     return _system_ids.find(system_id) != _system_ids.end();
