+++ cpp/src/mavsdk/core/mavlink_command_sender.hpp
@@ -43,7 +43,7 @@ public:
     using CommandResultCallback = std::function<void(Result, float)>;
 
     struct CommandInt {
-        uint8_t target_system_id{0};
+        uint32_t target_system_id{0};
         uint8_t target_component_id{0};
         MAV_FRAME frame = MAV_FRAME_GLOBAL_RELATIVE_ALT;
         uint16_t command{0};
@@ -62,7 +62,7 @@ public:
     };
 
     struct CommandLong {
-        uint8_t target_system_id{0};
+        uint32_t target_system_id{0};
         uint8_t target_component_id{0};
         uint16_t command{0};
         uint8_t confirmation = 0;
@@ -106,7 +106,7 @@ private:
         uint32_t maybe_param1{0}; // only for commands where this matters
         uint32_t maybe_param2{0}; // only for commands where this matters
         uint16_t command{0};
-        uint8_t target_system_id{0};
+        uint32_t target_system_id{0};
         uint8_t target_component_id{0};
 
         bool operator==(const CommandIdentification& other) const
